requirements.txt
aiosqlite==0.21.0
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
apprise==1.9.2
asgi-lifespan==2.1.0
asyncpg==0.30.0
attrs==25.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coolname==2.2.0
cryptography==44.0.2
dateparser==1.2.1
Deprecated==1.2.18
docker==7.1.0
duckdb==1.2.1
exceptiongroup==1.2.2
fastapi==0.115.11
fsspec==2025.2.0
graphviz==0.20.3
greenlet==3.1.1
griffe==1.6.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.12.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.3
oauthlib==3.2.2
opentelemetry-api==1.30.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pendulum==3.0.0
prefect==3.2.9
prometheus_client==0.21.1
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-socks==2.7.1
pytz==2025.1
PyYAML==6.0.2
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.23.1
ruamel.yaml==0.18.10
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
text-unidecode==1.3
time-machine==2.16.0
toml==0.10.2
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
ujson==5.10.0
urllib3==2.3.0
uv==0.6.4
uvicorn==0.34.0
websockets==15.0
wrapt==1.17.2
zipp==3.21.0
